Strategy code organized into functions not relying on context
