+++ lib/strategy.js
@@ -1,22 +1,12 @@
-// Module dependencies.
-var passport = require("passport-strategy"),
-  crypto = require("crypto"),
-  cose2jwk = require("cose-to-jwk"),
-  jwk2pem = require("jwk-to-pem"),
-  base64url = require("base64url"),
-  util = require("util"),
-  utils = require("./utils"),
-  Attestation = require("./fido2/attestation"),
-  AuthenticatorData = require("./fido2/authenticatordata"),
-  SessionStore = require("./store/session");
-
-// Constants for authenticator data flags.
-var USER_PRESENT = 0x01;
-var USER_VERIFIED = 0x04;
+const passport = require("passport-strategy");
+const base64url = require("base64url");
+const util = require("util");
+const utils = require("./utils");
+const SessionStore = require("./store/session");
+const operations = require("./strategy-operations");
 
 /**
  * Create a new `Strategy` object.
- *
  */
 function Strategy(options, verify, verifySignCount, register) {
   if (typeof options == "function") {
@@ -43,238 +33,50 @@ function Strategy(options, verify, verifySignCount, register) {
 // Inherit from `passport.Strategy`.
 util.inherits(Strategy, passport.Strategy);
 
+function validated(self, req, err, ok, ctx, clientData) {
+  if (err) {
+    return self.error(err);
+  }
+  if (!ok) {
+    return self.fail(ctx, 403);
+  }
+  ctx = ctx || {};
+
+  // Verify that the origin contained in client data matches the origin of this
+  // app (which is the relying party).
+  var origin = utils.origin(req);
+  if (origin !== clientData.origin) {
+    return self.fail({ message: "Origin mismatch" }, 403);
+  }
+
+  // TODO: Verify the state of Token Binding for the TLS connection over which
+  // the attestation was obtained.
+
+  if (clientData.type === "webauthn.get") {
+    return operations.validateCredentialsGet(self, req, ctx);
+  } else if (clientData.type === "webauthn.create") {
+    return operations.validateCredentialsCreate(self, req, ctx);
+  } else {
+    return self.fail(
+      { message: "Unsupported response type: " + clientData.type },
+      400,
+    );
+  }
+}
+
 Strategy.prototype.authenticate = function (req, options) {
+  var self = this;
+
   var response = req.body.response;
   var clientDataJSON = base64url.decode(response.clientDataJSON);
   var clientData = JSON.parse(clientDataJSON);
 
-  var self = this;
-
-  function validated(err, ok, ctx) {
-    if (err) {
-      return self.error(err);
-    }
-    if (!ok) {
-      return self.fail(ctx, 403);
-    }
-    ctx = ctx || {};
-
-    // Verify that the origin contained in client data matches the origin of this
-    // app (which is the relying party).
-    var origin = utils.origin(req);
-    if (origin !== clientData.origin) {
-      return self.fail({ message: "Origin mismatch" }, 403);
-    }
-
-    // TODO: Verify the state of Token Binding for the TLS connection over which
-    // the attestation was obtained.
-
-    if (clientData.type === "webauthn.get") {
-      // https://www.w3.org/TR/webauthn-2/#sctn-verifying-assertion
-
-      // TODO: Verify that credentials was in allowedCredentials, if set
-
-      var userHandle = null;
-      if (response.userHandle) {
-        userHandle = base64url.toBuffer(response.userHandle);
-      }
-
-      if (!ctx.user) {
-        // If the user was not identified before the authentication ceremony was
-        // initiated, response.userHandle must be present.
-        //
-        // NOTE: User handle being set should imply resident keys (???)
-        if (!userHandle) {
-          return self.fail({ message: "User handle not set" }, 403);
-        }
-      } else {
-        if (userHandle && Buffer.compare(ctx.user.id, userHandle) != 0) {
-          // If the user was identified before the authentication ceremony was
-          // initiated, if response.userHandle is present, it must map to the
-          // same user.
-          return self.fail(
-            { message: "User handle does not map to user" },
-            403,
-          );
-        }
-      }
-
-      var b_authenticatorData = base64url.toBuffer(response.authenticatorData);
-      var authenticatorData = AuthenticatorData.parse(b_authenticatorData);
-
-      // TODO: Support appID extension for rpIdHash
-
-      // Verify that the RP ID hash contained in authenticator data matches the
-      // hash of this app's (which is the relying party) RP ID.
-      if (!utils.checkRPIDHash(req, authenticatorData.rpIdHash)) {
-        return self.fail({ message: "RP ID hash mismatch" }, 403);
-      }
-
-      // Verify that the user present bit is set in authenticator data flags.
-      if (!(authenticatorData.flags & USER_PRESENT)) {
-        return self.fail({ message: "User not present" }, 403);
-      }
-
-      // TODO: Verify that extensions are as expected.
-
-      var id = req.body.id;
-      var flags = {
-        userPresent: !!(authenticatorData.flags & USER_PRESENT),
-        userVerified: !!(authenticatorData.flags & USER_VERIFIED),
-      };
-
-      const done = (err, user, publicKey, info) => {
-        utils.verificationDone(self, err, user, publicKey, info, {
-          clientDataJSON,
-          b_authenticatorData,
-          authenticatorData,
-          response,
-          id,
-        });
-      };
-
-      try {
-        if (self._passReqToCallback) {
-          var arity = self._verify.length;
-          switch (arity) {
-            case 5:
-              return self._verify(req, id, userHandle, flags, done);
-            default:
-              return self._verify(req, id, userHandle, done);
-          }
-        } else {
-          var arity = self._verify.length;
-          switch (arity) {
-            case 4:
-              return self._verify(id, userHandle, flags, done);
-            default:
-              return self._verify(id, userHandle, done);
-          }
-        }
-      } catch (ex) {
-        return self.error(ex);
-      }
-    } else if (clientData.type === "webauthn.create") {
-      // https://www.w3.org/TR/webauthn-2/#sctn-registering-a-new-credential
-
-      var b_attestation = base64url.toBuffer(response.attestationObject);
-      var attestation = Attestation.parse(b_attestation);
-      var authenticatorData = AuthenticatorData.parse(
-        attestation.authData,
-        true,
-        false,
-      );
-
-      // Verify that the RP ID hash contained in authenticator data matches the
-      // hash of this app's (which is the relying party) RP ID.
-      if (!utils.checkRPIDHash(req, authenticatorData.rpIdHash)) {
-        return self.fail({ message: "RP ID hash mismatch" }, 403);
-      }
-
-      // Verify that the user present bit is set in authenticator data flags.
-      if (!(authenticatorData.flags & USER_PRESENT)) {
-        return self.fail({ message: "User not present" }, 403);
-      }
-
-      // TODO: Verify alg is allowed
-
-      // TODO: Verify that extensions are as expected.
-
-      var format = self._attestationFormats[attestation.fmt];
-      if (!format) {
-        return self.fail(
-          { message: "Unsupported attestation format: " + attestation.fmt },
-          400,
-        );
-      }
-
-      // Verify that the attestation statement conveys a valid attestation signature.
-      var hash = crypto.createHash("sha256").update(clientDataJSON).digest();
-      var vAttestation;
-      try {
-        vAttestation = format.verify(
-          attestation.attStmt,
-          attestation.authData,
-          hash,
-        );
-      } catch (ex) {
-        return self.fail({ message: ex.message }, 400);
-      }
-
-      var credentialId = base64url.encode(
-        authenticatorData.attestedCredentialData.credentialId,
-      );
-      var jwk = cose2jwk(
-        authenticatorData.attestedCredentialData.credentialPublicKey,
-      );
-      var pem = jwk2pem(jwk);
-      var flags = {
-        userPresent: !!(authenticatorData.flags & USER_PRESENT),
-        userVerified: !!(authenticatorData.flags & USER_VERIFIED),
-      };
-
-      const done = (err, user, info) => {
-        return utils.registrationDone(self, err, user, info);
-      };
-
-      try {
-        if (self._passReqToCallback) {
-          // TODO
-          //self._verify(req, username, password, verified);
-        } else {
-          var arity = self._register.length;
-          switch (arity) {
-            case 8:
-              return self._register(
-                ctx.user,
-                credentialId,
-                pem,
-                flags,
-                authenticatorData.signCount,
-                response.transports,
-                vAttestation,
-                done,
-              );
-            case 7:
-              return self._register(
-                ctx.user,
-                credentialId,
-                pem,
-                flags,
-                authenticatorData.signCount,
-                response.transports,
-                done,
-              );
-            case 6:
-              return self._register(
-                ctx.user,
-                credentialId,
-                pem,
-                flags,
-                authenticatorData.signCount,
-                done,
-              );
-            case 5:
-              return self._register(ctx.user, credentialId, pem, flags, done);
-            default:
-              return self._register(ctx.user, credentialId, pem, done);
-          }
-        }
-      } catch (ex) {
-        return self.error(ex);
-      }
-    } else {
-      return self.fail(
-        { message: "Unsupported response type: " + clientData.type },
-        400,
-      );
-    }
-  }
-
-  // Verify that the challenge (aka nonce) received from the client equals the
-  // challenge sent.
   var challenge = base64url.toBuffer(clientData.challenge);
-  this._store.verify(req, challenge, validated);
+
+  // Verify that the challenge (aka nonce) received from the client equals the challenge sent
+  this._store.verify(req, challenge, (err, ok, ctx) => {
+    validated(self, req, err, ok, ctx, clientData);
+  });
 };
 
 module.exports = Strategy;
